Expose Layout::clear
