+++ parley/src/tests/mod.rs
@@ -3,4 +3,5 @@
 
 mod test_analysis;
 mod test_builders;
+mod test_layout;
 mod utils;
