+++ parley/src/tests/test_builders.rs
@@ -3,72 +3,21 @@
 
 //! Test that the various builders produce the same results.
 
-use std::{borrow::Cow, path::PathBuf, sync::Arc, vec, vec::Vec};
+use std::{vec, vec::Vec};
 
-use fontique::{Collection, CollectionOptions, FontStyle, FontWeight, FontWidth, SourceCache};
-use parlance::FontFamilyName;
-use peniko::{Blob, color::palette};
+use fontique::{FontStyle, FontWeight, FontWidth};
+use peniko::color::palette;
 
-use super::utils::{ColorBrush, asserts::assert_eq_layout_data};
+use super::utils::{
+    ColorBrush,
+    asserts::assert_eq_layout_data,
+    fonts::{FONT_FAMILY_LIST, create_font_context},
+};
 use crate::{
     BaseDirection, BreakReason, FontContext, FontFamily, FontFeatures, FontVariations, Layout,
     LayoutContext, LineHeight, OverflowWrap, RangedBuilder, StyleProperty, StyleRunBuilder,
     TextStyle, TextWrapMode, TreeBuilder, WordBreak,
 };
-
-// TODO: `FONT_FAMILY_LIST`, `load_fonts`, and `create_font_context` are
-// duplicated between this crate and `parley_test`. We can't move the builder
-// tests into `parley_test` because they use private APIs, but should eventually
-// figure out some way to reduce the duplication.
-const FONT_FAMILY_LIST: &[FontFamilyName<'_>] = &[
-    FontFamilyName::Named(Cow::Borrowed("Roboto")),
-    FontFamilyName::Named(Cow::Borrowed("Noto Kufi Arabic")),
-];
-
-pub(crate) fn load_fonts(
-    collection: &mut Collection,
-    font_dirs: impl Iterator<Item = PathBuf>,
-) -> std::io::Result<()> {
-    for dir in font_dirs {
-        let paths = std::fs::read_dir(dir)?;
-        for entry in paths {
-            let entry = entry?;
-            if !entry.metadata()?.is_file() {
-                continue;
-            }
-            let path = entry.path();
-            if path
-                .extension()
-                .and_then(|ext| ext.to_str())
-                .is_none_or(|ext| !["ttf", "otf", "ttc", "otc"].contains(&ext))
-            {
-                continue;
-            }
-            let font_data = std::fs::read(&path)?;
-            collection.register_fonts(Blob::new(Arc::new(font_data)), None);
-        }
-    }
-    Ok(())
-}
-
-fn create_font_context() -> FontContext {
-    let mut collection = Collection::new(CollectionOptions {
-        shared: false,
-        system_fonts: false,
-    });
-    load_fonts(&mut collection, parley_dev::font_dirs()).unwrap();
-    for font in FONT_FAMILY_LIST {
-        if let FontFamilyName::Named(font_name) = font {
-            collection
-                .family_id(font_name)
-                .unwrap_or_else(|| panic!("{font_name} font not found"));
-        }
-    }
-    FontContext {
-        collection,
-        source_cache: SourceCache::default(),
-    }
-}
 
 /// Set of options for [`build_layout_with_ranged`].
 struct RangedOptions<'a> {
